PYI-775 - Refactored credential-issuer middleware to be shared to avoid duplicating logic for Journey and Debug routes.
