+++ src/app/credential-issuer/middleware.js
@@ -2,66 +2,10 @@ const axios = require("axios");
 const {
   API_BASE_URL,
   API_REQUEST_EVIDENCE_PATH,
-  API_SHARED_ATTRIBUTES_JWT_PATH,
-  SHARED_ATTRIBUTES_JWT_SIZE_LIMIT,
   EXTERNAL_WEBSITE_HOST,
 } = require("../../lib/config");
 
 module.exports = {
-  getSharedAttributesJwt: async (req, res, next) =>  {
-    try {
-      const apiResponse = await axios.get(
-        `${API_BASE_URL}${API_SHARED_ATTRIBUTES_JWT_PATH}`,
-        {
-          headers: {
-            "ipv-session-id": req.session.ipvSessionId
-          }
-        }
-      );
-
-      const sharedAttributesJwt = apiResponse?.data;
-      const base64regex = /^[A-Za-z0-9-_=]+\.[A-Za-z0-9-_=]+\.?[A-Za-z0-9-_.+/=]*$/;
-
-      if (!sharedAttributesJwt) {
-        res.status(500);
-        return res.send("Missing JWT");
-      } else if (sharedAttributesJwt.length > SHARED_ATTRIBUTES_JWT_SIZE_LIMIT) {
-        res.status(500);
-        return res.send("JWT exceeds maximum limit");
-      } else if (!base64regex.test(sharedAttributesJwt)) {
-        res.status(500);
-        return res.send("Invalid base64 encoded JWT");
-      }
-
-      req.session.sharedAttributesJwt = sharedAttributesJwt;
-
-      next();
-    } catch (e) {
-      next(e);
-    }
-  },
-
-  buildCredentialIssuerRedirectURL: async (req, res, next) => {
-    const cri = req.session?.criConfig?.find(criConfig => criConfig.id === req.query.id);
-
-    if (!cri) {
-      res.status(500);
-      return res.send("Could not find configured CRI");
-    }
-
-    req.redirectURL = new URL(cri.authorizeUrl);
-    req.redirectURL.searchParams.append("response_type", "code");
-    req.redirectURL.searchParams.append("client_id", cri.ipvClientId);
-    req.redirectURL.searchParams.append("state", "test-state");
-    req.redirectURL.searchParams.append("redirect_uri", `${EXTERNAL_WEBSITE_HOST}/credential-issuer/callback?id=${cri.id}`);
-    req.redirectURL.searchParams.append("request", req.session.sharedAttributesJwt);
-
-    next();
-  },
-  redirectToAuthorize: async (req, res) => {
-    res.redirect(req.redirectURL);
-  },
-
   addCallbackParamsToRequest: async (req, res, next) => {
     req.credentialIssuer = {};
 
@@ -93,7 +37,7 @@ module.exports = {
       res.status = apiResponse?.status;
       next();
     } catch (error) {
-      if (error?.response?.status == 404) {
+      if (error?.response?.status === 404) {
         res.status = error.response.status;
       } else {
         res.error = error.name;
@@ -101,7 +45,9 @@ module.exports = {
       next(error);
     }
   },
-
+  redirectToAuthorize: async (req, res) => {
+    res.redirect(req.redirectURL);
+  },
   redirectToDebugPage: async (req, res) => {
     res.redirect("/debug/");
   },
