+++ src/app/journey/middleware.test.js
@@ -10,12 +10,29 @@ describe("journey middleware", () => {
   const axiosStub = {};
   const configStub = {
     API_BASE_URL: "https://example.org/subpath",
+    API_SHARED_ATTRIBUTES_JWT_PATH: '/jwtPath',
+    EXTERNAL_WEBSITE_HOST: "https://callbackaddres.org"
   };
+
+  const sharedAttributeHelper = proxyquire("../shared/sharedAttributeHelper", {
+    axios: axiosStub,
+    "../../lib/config": configStub,
+  });
+
+  const sharedCriHelper = proxyquire("../shared/criHelper", {
+    axios: axiosStub,
+    "../../lib/config": configStub,
+  });
 
   const middleware = proxyquire("./middleware", {
     axios: axiosStub,
     "../../lib/config": configStub,
+    "../shared/sharedAttributeHelper": sharedAttributeHelper,
+    "../shared/../shared/criHelper": sharedCriHelper,
   });
+
+
+
 
   beforeEach(() => {
     res = {
@@ -110,7 +127,7 @@ describe("journey middleware", () => {
     beforeEach(() => {
       eventResponses = [
         {
-          data: { redirect: { cri: { id: 'someid', authorizeUrl: authorizeUrl, request: '' } } }
+          data: { redirect: { cri: { id: 'someid', authorizeUrl: authorizeUrl, request: 'req', ipvClientId: "clientId" } } }
         },
       ];
       req = {
@@ -124,11 +141,19 @@ describe("journey middleware", () => {
       eventResponses.forEach((er, index) => {
         callBack.onCall(index).returns(eventResponses[index]);
       });
+
+      const getStub = sinon.stub();
+      getStub.withArgs(`${configStub.API_BASE_URL}${configStub.API_SHARED_ATTRIBUTES_JWT_PATH}`,
+        {
+          headers : {"ipv-session-id": req.session.ipvSessionId}
+        })
+        .returns({data: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJkYXRlT2ZCaXJ0aHMiOltdLCJhZGRyZXNzZXMiOltdLCJuYW1lcyI6W10sImFkZHJlc3NIaXN0b3J5IjpbXX0.DwQQOldmOYQ1Lv6OJETzks7xv1fM7VzW0O01H3-uQqQ_rSkCZrd2KwQHHzo0Ddw2K_LreePy-tEr-tiPgi8Yl604n3rwQy6xBat8mb4lTtNnOxsUOYviYQxC5aamsvBAS27G43wFejearXHWzEqhJhIFdGE4zJkgZAKpLGzvOXLvX4NZM4aI4c6jMgpktkvvFey-O0rI5ePh5RU4BjbG_hvByKNlLr7pzIlsS-Q8KuIPawqFJxN2e3xfj1Ogr8zO0hOeDCA5dLDie78sPd8ph0l5LOOcGZskd-WD74TM6XeinVpyTfN7esYBnIZL-p-qULr9CUVIPCMxn-8VTj3SOw=='})
+      axiosStub.get = getStub;
     });
 
-    it("should be redirected to the url supplied from authorizeUrl", async function() {
+    it("should be redirected to a valid redirectURL", async function() {
       await middleware.updateJourneyState(req, res, next);
-      expect(res.redirect).to.have.been.calledWith(authorizeUrl);
+      expect(req.redirectURL.toString()).to.equal("https://someurl.com/?response_type=code&client_id=clientId&state=test-state&redirect_uri=https%3A%2F%2Fcallbackaddres.org%2Fcredential-issuer%2Fcallback%3Fid%3Dsomeid&request=eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJkYXRlT2ZCaXJ0aHMiOltdLCJhZGRyZXNzZXMiOltdLCJuYW1lcyI6W10sImFkZHJlc3NIaXN0b3J5IjpbXX0.DwQQOldmOYQ1Lv6OJETzks7xv1fM7VzW0O01H3-uQqQ_rSkCZrd2KwQHHzo0Ddw2K_LreePy-tEr-tiPgi8Yl604n3rwQy6xBat8mb4lTtNnOxsUOYviYQxC5aamsvBAS27G43wFejearXHWzEqhJhIFdGE4zJkgZAKpLGzvOXLvX4NZM4aI4c6jMgpktkvvFey-O0rI5ePh5RU4BjbG_hvByKNlLr7pzIlsS-Q8KuIPawqFJxN2e3xfj1Ogr8zO0hOeDCA5dLDie78sPd8ph0l5LOOcGZskd-WD74TM6XeinVpyTfN7esYBnIZL-p-qULr9CUVIPCMxn-8VTj3SOw%3D%3D");
     });
 
     it("should raise an error when missing authorizeUrl", async () => {
@@ -145,6 +170,8 @@ describe("journey middleware", () => {
         eventResponses.forEach((er, index) => {
           callBack.onCall(index).returns(eventResponses[index]);
         });
+
+
       });
 
       await middleware.handleJourneyPage(req, res, next);
